do  not run any migration if project == :eog for now fine tuning follows
Former-commit-id: c5b134c76ca3477245d10c2e110e246278d19763

+++ db/migrate/20100312111727_add_project_id_to_collections.rb
@@ -1,17 +1,21 @@
 class AddProjectIdToCollections < ActiveRecord::Migration
 
   def self.up
+  unless Project.name.to_sym == :eog
 
     change_table :collections do |t|
       t.string :project_id
     end
 
   end
+  end
 
   def self.down
+  unless Project.name.to_sym == :eog
 
     remove_column :collections, :project_id
 
+  end
   end
 
 end
