+++ db/migrate/20100609170100_create_text_materials.rb
@@ -1,5 +1,6 @@
 class CreateTextMaterials < ActiveRecord::Migration
   def self.up
+  unless Project.name.to_sym == :eog
     create_table :text_materials do |t|
       t.references :interview
       t.string :document_type
@@ -8,8 +9,11 @@ class CreateTextMaterials < ActiveRecord::Migration
       t.integer :document_file_size
     end
   end
+  end
 
   def self.down
+  unless Project.name.to_sym == :eog
     drop_table :text_materials
+  end
   end
 end
