+++ db/migrate/20131127174503_remove_obsolete_columns_from_segments.rb
@@ -1,10 +1,14 @@
 class RemoveObsoleteColumnsFromSegments < ActiveRecord::Migration
   def self.up
+  unless Project.name.to_sym == :eog
     remove_column :segments, :previous_segment_id, :following_segment_id
+  end
   end
 
   def self.down
+  unless Project.name.to_sym == :eog
     add_column :segments, :previous_segment_id, :integer
     add_column :segments, :following_segment_id, :integer
+  end
   end
 end
